# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "TaskRabbit Clone API",
      "version": "1.0.0",
      "description": "API documentation for the TaskRabbit clone backend"
    },
    "basePath": "/api",
    "schemes": [
      "https"
    ],
    "securityDefinitions": {
      "JWT": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/admin/users/{userId}": {
        "get": {
          "tags": [
            "Admin"
          ],
          "summary": "Get detailed information about a specific user by their ID",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "User retrieved successfully"
            },
            "401": {
              "description": "Authentication required"
            },
            "403": {
              "description": "Access denied"
            },
            "404": {
              "description": "User with the specified ID not found"
            }
          }
        },
        "put": {
          "tags": [
            "Admin"
          ],
          "summary": "Update the details of a specific user",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/UpdateUserRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User updated successfully"
            },
            "401": {
              "description": "Authentication required"
            },
            "403": {
              "description": "Access denied"
            },
            "404": {
              "description": "User with the specified ID not found"
            }
          }
        },
        "delete": {
            "tags": [
              "Admin"
            ],
            "summary": "Delete a specific user",
            "parameters": [
              {
                "name": "userId",
                "in": "path",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "204": {
                "description": "User deleted successfully"
              },
              "401": {
                "description": "Authentication required"
              },
              "403": {
                "description": "Access denied"
              },
              "404": {
                "description": "User with the specified ID not found"
              }
            }
          }
        }
      },
      "definitions": {
...
        "UpdateUserRequest": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        }
      }
    }